fix: Fix failing tests
