+++ src/desktop/apps/clear_cache/index.coffee
@@ -1,5 +1,5 @@
 express = require 'express'
-{ flushall } = require '../../lib/cache'
+{ flushall } = require('../../lib/cache').cache
 
 app = module.exports = express()
 app.set 'views', __dirname
