+++ src/lib/cacheClient.tsx
@@ -1,5 +1,5 @@
 import createLogger from "v2/Utils/logger"
-import artsyCache from "./cache"
+import { cache as artsyCache } from "./cache"
 interface RedisCache {
   get: (key: string) => Promise<string>
   set: (
