+++ src/test/acceptance/helpers/index.js
@@ -111,8 +111,7 @@ const warn = e => console.log("warning", chalk.yellow(e))
 
 const startForce = port =>
   new Promise((resolve, reject) => {
-    const app = require("../../../common-app")
-    app.initialize(() => {})
+    const { app } = require("../../../index")
     servers.push(
       app.listen(port, err => {
         if (err) reject(err)
